third_party/blink/web_tests/fast/dom/shadow/distribution-for-detached-subtree-expected.txt

CONSOLE WARNING: line 8: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
distribution should be recalculated correctly for a detached subtree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS nodeList.length is expectedNodes.length
PASS nodeList.length is expectedNodes.length
PASS nodeList.item(0) is expectedNodes[0]
PASS successfullyParsed is true

TEST COMPLETE
